scraper/requirements.txt
# Core dependencies
fastapi>=0.115.12
uvicorn>=0.34.3
pydantic>=1.8.0
python-multipart>=0.0.20
jinja2>=3.1.6
python-dotenv>=1.1.0

# Database
sqlalchemy>=2.0.41
sqlalchemy-json>=0.7.0
psycopg2-binary>=2.9.10
alembic>=1.16.2

# Web scraping
requests>=2.32.4
scrapy>=2.13.2
scrapy-fake-useragent>=1.4.4
scrapy-playwright>=0.0.43
sentence-transformers>=4.1.0
scikit-learn>=1.3.0
gliner>=0.2.21
playwright>=1.52.0

# Image processing
pillow>=11.2.1
imagehash>=4.3.2

# Caching and sessions
redis>=6.2.0
cachetools>=6.1.0

# Elasticsearch
elasticsearch>=8.12.0
elasticsearch-dsl>=8.12.0

# Auth
PyJWT>=2.10.1

# Utilities
pyyaml>=6.0.2
numpy>=2.2.6
aiohttp>=3.12.13
websockets>=15.0.1
watchdog>=6.0.0 

